removed search page
