+++ src/scss/_search-syntax-highlighting.scss
@@ -19,6 +19,7 @@
   left: 0;
   position: absolute;
   top: 0;
+  width: 100%;
   z-index: 0;
 }
