+++ src/services/advancedSearchQuery/readme.md
@@ -9,7 +9,7 @@ The parser will build an abstract syntax tree AST with the grammar elements (PHP
 After the AST is build successfully (syntax errors are reported to the user) it can be visited. Here the visitor pattern is used and there are tree visitors located in `src/services/advancedSearchQuery/visitors`: `DepthValidatorVisitor`, `FieldValidatorVisitor`, and `QueryBuilderVisitor`.
 
 1. `DepthValidatorVisitor` checks the complexity of the query. How deep it the AST? (Not used currently.)
-2. `FieldValidatorVisitor` checks if fields/values are allowed, e.g., the `timestamped` field is only meaningful for experiments. Feedback is given for illegal fields/values.
+2. `FieldValidatorVisitor` checks if fields/values are allowed, e.g., the `timestamped` field is only meaningful for experiments or items. Feedback is given for illegal fields/values.
 3. `QueryBuilderVisitor` is the workhorse and builds the SQL where clause.
 
 The `FieldValidatorVisitor` and `QueryBuilderVisitor` use collectors (`src/services/advancedSearchQuery/collectors`) to collect all the invalid fields and the where clause, respectively.
