+++ src/sql/schema171.sql
@@ -3,6 +3,10 @@ ALTER TABLE `experiments` ADD `canread_is_immutable` TINYINT UNSIGNED NOT NULL D
 ALTER TABLE `experiments` ADD `canwrite_is_immutable` TINYINT UNSIGNED NOT NULL DEFAULT 0;
 ALTER TABLE `experiments_templates` ADD `canread_is_immutable` TINYINT UNSIGNED NOT NULL DEFAULT 0;
 ALTER TABLE `experiments_templates` ADD `canwrite_is_immutable` TINYINT UNSIGNED NOT NULL DEFAULT 0;
+ALTER TABLE `experiments_templates` ADD `date` date NULL DEFAULT NULL;
+ALTER TABLE `experiments_templates` ADD `elabid` VARCHAR(255) NULL DEFAULT NULL;
+ALTER TABLE `items_types` ADD `date` date NULL DEFAULT NULL;
+ALTER TABLE `items_types` ADD `elabid` VARCHAR(255) NULL DEFAULT NULL;
 ALTER TABLE `items` ADD `canread_is_immutable` TINYINT UNSIGNED NOT NULL DEFAULT 0;
 ALTER TABLE `items` ADD `canwrite_is_immutable` TINYINT UNSIGNED NOT NULL DEFAULT 0;
 ALTER TABLE `items_types` ADD `canread_is_immutable` TINYINT UNSIGNED NOT NULL DEFAULT 0;
