+++ src/sql/structure.sql
@@ -307,6 +307,8 @@ CREATE TABLE `experiments_templates` (
   `body` text,
   `category` INT UNSIGNED NULL DEFAULT NULL,
   `custom_id` INT UNSIGNED NULL DEFAULT NULL,
+  `date` date NULL DEFAULT NULL,
+  `elabid` varchar(255) NULL DEFAULT NULL,
   `title` varchar(255) NOT NULL,
   `userid` int(10) UNSIGNED DEFAULT NULL,
   `locked` tinyint UNSIGNED NOT NULL DEFAULT 0,
@@ -722,6 +724,8 @@ CREATE TABLE `items_types` (
   `userid` INT UNSIGNED NOT NULL,
   `team` int(10) UNSIGNED NOT NULL,
   `title` varchar(255) NOT NULL,
+  `date` date NULL DEFAULT NULL,
+  `elabid` varchar(255) NULL DEFAULT NULL,
   `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `color` varchar(6) DEFAULT '29aeb9',
   `custom_id` INT UNSIGNED NULL DEFAULT NULL,
