+++ src/templates/search-syntax-doc-modal.html
@@ -2,15 +2,46 @@
   <div class='modal-dialog modal-xl' role='document'>
     <div class='modal-content'>
       <div class='modal-header'>
-        <h5 class='modal-title' id='advancedSearchModalLabel'>{{ 'Search Syntax'|trans }}</h5>
+        <h5 class='modal-title' id='advancedSearchModalLabel'>{{ 'Search help'|trans }}</h5>
         <button type='button' class='close' data-dismiss='modal' aria-label='{{ 'Close'|trans }}'>
           <span aria-hidden='true'>&times;</span>
         </button>
       </div>
       <div class='modal-body' data-wait='{{ 'Please wait…'|trans }}'>
-        <h3>Search terms</h3>
-        <p>The search is performed against title, body, date, and elabid.<br>
-            The entire team will be searched unless there is an author field.</p>
+        <h3>{{ 'Simple search'|trans }}</h3>
+        <p>{{ 'When searching a term in the search bar, results will contain this term in the title, main text, date or elabid.'|trans }}</p>
+        <h3>{{ 'Advanced search'|trans }}</h3>
+        {% set fieldValue = "<code><span class='token keyword'>%s</span><span class='token punctuation'>:</span><span class='token string'>%s</span></code>"|format('field'|trans, 'value'|trans) %}
+        <p>{{ 'The search bar supports specific search such as "title:Something" and "catogory:Tests" (%s), that you can combine with logical expressions (AND, OR, NOT...) and group with parenthesis.'|trans|format(fieldValue)|raw }}</p>
+        <p>{{ 'To search in extra fields (metadata), use the dedicated helper to build your query. For example, to search all entries where "Temperature" is 37, you would use: "extrafield:Temperature:37" in the search bar.'|trans }}</p>
+        <h3>{{ 'Search terms'|trans }}</h3>
+        <p>{{ 'Allowed Fields'|trans }} :</p>
+        <div class='row'>
+          <div class='col-md-6'>
+            <ul class='list-unstyled'>
+              <li class='mb-2'><kbd>attachment:</kbd> {{ 'yes, no, simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>body:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>category:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>custom_id:</kbd> {{ 'number'|trans }}</li>
+              <li class='mb-2'><kbd>date:</kbd> {{ 'see date format'|trans }}</li>
+              <li class='mb-2'><kbd>elabid:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>extrafield:key:value</kbd> {{ 'simple or quoted term'|trans }}</li>
+            </ul>
+          </div>
+
+          <div class='col-md-6'>
+            <ul class='list-unstyled'>
+              <li class='mb-2'><kbd>group:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>locked:</kbd> {{ 'yes or no'|trans }}</li>
+              <li class='mb-2'><kbd>owner:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>rating:</kbd> {{ '1, 2, 3, 4, 5, or unrated or 0'|trans }}</li>
+              <li class='mb-2'><kbd>status:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>timestamped:</kbd> {{ 'yes or no'|trans }}</li>
+              <li class='mb-2'><kbd>title:</kbd> {{ 'simple or quoted term'|trans }}</li>
+              <li class='mb-2'><kbd>visibility:</kbd> {{ 'simple or quoted term'|trans }}</li>
+            </ul>
+          </div>
+        </div>
         <dl>
           <dt>Term</dt>
           <dd>A sequence of characters without space (<code>&blank;</code>),
